Use plan json instead of jsonpickle.
jsonpickle apparently sometimes encodes numpy types in a way which
cannot be decoded. I'm not sure exactly why this is; I think it may have
to do with nested ndarrays of types which are not plain python types. In
particular it screws up np.float64(0.7), for example.

This is obviously a problem; we want to be sure that our serialised type
can be deserialised, or it is not much use! Get around this by using
plain json and manually converting numpy arrays to lists and types to
string tuples. It's a hack, but it should work.